service/requirements.txt
click~=7.0 
psycopg2-binary~=2.8.3 
requests2~=2.16.0
